requirements.txt
alembic==1.7.7
Authlib==1.0.1
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///Users/ktietz/Code/oss/ci_pkgs/cffi_1624487777258/work
chardet @ file:///Users/ktietz/Code/oss/ci_pkgs/chardet_1626783467640/work
charset-normalizer==2.0.12
click==8.1.2
conda==4.10.1
conda-package-handling @ file:///Users/ktietz/Code/oss/ci_pkgs/conda-package-handling_1626784934972/work
cryptography @ file:///Users/ktietz/Code/oss/ci_pkgs/cryptography_1624488650114/work
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///Users/ktietz/Code/oss/ci_pkgs/pysocks_1626781349491/work
python-dotenv==0.20.0
requests==2.27.1
ruamel-yaml-conda @ file:///Users/ktietz/Code/oss/ci_pkgs/ruamel_yaml_1626786937887/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
SQLAlchemy==1.4.35
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
visitor==0.1.3
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
